Add column 'workflow_state' to WorkflowState based models
Use column value instead of calculating value based on related cycle tasks states each time when model json is returned
Update workflow_state column for related objects when cycle task state is changed
