+++ src/ggrc/models/account_balance.py
@@ -31,6 +31,7 @@ class AccountBalance(
     mixins.base.ContextRBAC,
     mixins.ScopeObject,
     mixins.Folderable,
+    mixins.WithWorkflowState,
     fulltext_mixin.Indexed,
     db.Model,
 ):
