+++ src/ggrc/models/assessment.py
@@ -27,6 +27,7 @@ from ggrc.models.mixins import Notifiable
 from ggrc.models.mixins import TestPlanned
 from ggrc.models.mixins import LastDeprecatedTimeboxed
 from ggrc.models.mixins import VerifiedDate
+from ggrc.models.mixins import WithWorkflowState
 from ggrc.models.mixins import reminderable
 from ggrc.models.mixins import statusable
 from ggrc.models.mixins import labeled
@@ -70,6 +71,7 @@ class Assessment(Assignable,
                  BusinessObject,
                  with_sox_302.WithSOX302FlowReadOnly,
                  WithCustomRestrictions,
+                 WithWorkflowState,
                  Indexed,
                  db.Model):
   """Class representing Assessment.
