+++ src/ggrc/models/data_asset.py
@@ -25,6 +25,7 @@ class DataAsset(Roleable,
                 mixins.base.ContextRBAC,
                 mixins.Folderable,
                 mixins.ScopeObject,
+                mixins.WithWorkflowState,
                 Indexed,
                 db.Model):
   """Class representing DataAsset."""
