+++ src/ggrc/models/product.py
@@ -32,6 +32,7 @@ class Product(Roleable,
               mixins.base.ContextRBAC,
               mixins.ScopeObject,
               mixins.Folderable,
+              mixins.WithWorkflowState,
               Indexed,
               db.Model):
   """Representation for Product model."""
