+++ src/ggrc/models/program.py
@@ -40,6 +40,7 @@ class Program(mega.Mega,
               mixins.BusinessObject,
               proposal.Proposalable,
               mixins.Folderable,
+              mixins.WithWorkflowState,
               ft_mixin.Indexed,
               db.Model):
   """Representation for Program model."""
