+++ src/ggrc/models/project.py
@@ -24,6 +24,7 @@ class Project(Roleable,
               mixins.base.ContextRBAC,
               mixins.ScopeObject,
               mixins.Folderable,
+              mixins.WithWorkflowState,
               Indexed,
               db.Model):
   """Representation for Project model."""
