+++ src/ggrc/models/requirement.py
@@ -32,6 +32,7 @@ class Requirement(Roleable,
                   mixins.base.ContextRBAC,
                   mixins.BusinessObject,
                   mixins.Folderable,
+                  mixins.WithWorkflowState,
                   Indexed,
                   db.Model):
   """Requirement model."""
