+++ src/ggrc/models/technology_environment.py
@@ -26,6 +26,7 @@ class TechnologyEnvironment(mixins.CustomAttributable,
                             mixins.base.ContextRBAC,
                             mixins.ScopeObject,
                             mixins.Folderable,
+                            mixins.WithWorkflowState,
                             db.Model,
                             Indexed):
   """Representation for TechnologyEnvironment model."""
