+++ src/ggrc/models/threat.py
@@ -25,6 +25,7 @@ class Threat(Roleable,
              mixins.base.ContextRBAC,
              mixins.BusinessObject,
              mixins.Folderable,
+             mixins.WithWorkflowState,
              Indexed,
              db.Model):
   __tablename__ = 'threats'
