+++ src/ggrc/models/vendor.py
@@ -27,6 +27,7 @@ class Vendor(Roleable,
              mixins.base.ContextRBAC,
              mixins.ScopeObject,
              mixins.Folderable,
+             mixins.WithWorkflowState,
              Indexed,
              db.Model):
   """Class for Vendor model"""
